saving game
Did a ton of work on saving the game. Most game objects can be serialized into JSON (well… they WILL be). Got a lot of the way on creating the save screen that shows the saved games. What a PIA working with files is. Must create a lib for this shiz after all my work. Anyway, have to wire the save screen up to the LevelView and see if it actually works. I really should mock it out more. This game is becoming a ton of work. I'm really enjoying myself, but I worry like a lot of unfinished projects in my past if I'm making this too complicated. The gameplay is UBER simple so I fail to see how I could make it simpler, nor the code more spaghetti-ish to git-r-done faster. Maybe it's just this genre of game takes a lot of work.